Bug fix: valgrind reports unconditional jumps
Trigger: When the node have some unassigned pipe end-point with some
accessor defined, read/write accessor with pstd_type will cause valgrind error.

Explain: Although for all the accesssor we have one type callback
initialize it, but when the pipe is unassigned, the type system will
ignore the pipe and the accessor won't be initialized in this case.
However, the read/write function in pstd_type won't check the
initialzation flag of the accessor data structure, which case the
valgrind error.

Fix:
1. Make the read/write function check the init flag of the accessor
2. Keep tracking the init state pipe type info data and do not work on
the uninitialized type
3. Make sure the accessor read works with zero/unassigned inputs