Bug fix: segv fault when the async_init fails
Cause: orignally sched_async_post_task disposes the async_init task when
it's done. However, this makes the sched_step_next unable to set the
signal pipe. This causes the crash.

Fix: We redefined the behavior of the sched_async_post_task funciton,
which makes the sched_task_t::exec_task always valid after the function
is called. So the scheduler will be able to set the signal pipe