Bug fix: The dispatcher becomes ineffecient when the IO event fail to dispatch
Previously we only allows task event added to pending list, this makes
in some case, the IO event has been poped up, but failed to dispatch.
At this point, it blocks other task event to be dispatched, which means
the running task can not be resolved, which makes the scheduler
saturate. This causes dead lock

TODO: We still don't understand why it's possible a IO event cannot be
dispatched because of the saturation limit, since we update the event
mask, only when there's at least on slot, an IO event can be poped up