Bug fix: PSTD file cache stat cache may be inconsistent with the data buffer
Cannot actually reproduce the bug, since it's hard to repo. The static
file server returns a bad HTTP response because the pstd_file_size
returns an different size. It seems the size is the latest the size but
the data buffer is old.

The most suspicious cuase of this is: When the pstd_fcache_open is
called with a uncached file, the modification timestamp is T second.
However, when the cache entry get the timestamp, it's in T + 1 second.
During the time between stat and read, the target file has been updated
to a different size. Thus when the file cache is expired, it call the
stat again. However the last modification has timestamp T. Thus the
cache entry still can be used.

The root cause of this is the timestamp should be before the actual file
access, so that we can make sure if the timestamp is strictly greater than
the file modification timestamp, the file must be untouched during time.

However, the previous implementaion actually access the stat data first
and then ask for the timestamp. This is unlikely to happen, because the
time window to trigger this bug is tiny. But ideally it can happen. And
the bug that causes the server returns a bad request is one instance of
this bug.