Fix the pthread errno convention misuse
Previously we assume the pthread functions as standard POSIX error code
convention, which means < 0 indicates error and errno will be used.

However pthread uses a different convention != 0 indicates error and the
errno will be returned as the return value. Thus we need to change all
the pthread functions.

Previously:
	if(pthread_func() < 0)

Now:
	if((errno = pthread_func()) != 0)