Add round robin move threshold for the dispatcher for performance
Issue: All the performance evaluation indicates that the maximum QPS
cannot be larger than 500k, and dispatcher is the bottleneck. This is
concerning, since for HTTP request this is probably Ok for a lot of
cases. However, in other scenario, this is a huge limit.

Cause: When the worker thread consumes the event too fast (This can
caused by either the task is simple, or we have too many worker thread
running). The job queue for each worker will be constantly empty
(Dispatched one is consumed before the next one is enqueue). This makes
the dispatcher has to notify the waiting worker multiple times.
Form the worker side, once it gets a job to do, it will be done shortly
and goes back to sleep. This makes the dispatcher has to spend most of
the performance on signaling the worker thread for new jobs. This
invovles system call, thus it impact the performance of the dispatcher
largely. The system throughput is limited by this part in this scenario
as well.

Solution: This makes us think about the nature of the dispatcher. Not in
all the cases the perfect load balancing is helpful. For example, if the
worker threads' computation power is much higher than it required in all
cases (especially in the high pressure environment, which means each of
the event doesn't take a long time to process or this is not a CPU bound
operation). It's not optimal to distribute the job into different
workers, since it reduces the performance of the dispatcher, which is
the bottleneck in this scenario.
Thus we intorduce a configuration called
scheduler.worker.round_robin_move_threshold.
If the current worker's job queue is not larger than the threshold, the
round robin load balancer will keep sending jobs to current worker
instead of the next one. For the case that the job is processed faster
than it's dispatched case, the dispatcher will pickup the same worker
util the queue has grown to a larger size. (In this case we should
adjust the threshold to larger value for example 10) In this case we
don't have the issue of waste computing power, because the power is more
than need.

However, in a CPU bound case which means the worker's job takes longer
time than dispatcher to dispatch the job. In this case, the perfect load
balancing is helpful, otherwise we are wasting the computing power. So
in this case the threshold needs to be set to 0.