Bug fix: After use _sched_waiting flags the async task hangs
Trigger: When an async task involved in the event queue, it's possible
that the dispatcher cannot be waked up sucessfully after the event
raised.

Cause: Since we only call pthread_cond_signal at most once in the code.
But previoosly we assume every event look will be able to wake up the
dispatcher. However, with the async task, this is on longer true. The
dispatcher might reject the async event temporarily. However, once an
async task decide to send the signal, the dispatcher won't be waken up
by it. However the waiting flag is cleared in this case.

Solution: Also check if the event look can wake up the dispatcher before
actually signal.