Bug fix: The dispatcher fail to wake up when pressure is extremely high
Cause: Since we use mask as the waiting flag in prevoius change. But
When the dispatcher has all the worker saturated, the scheduler needs to
reject the IO event. In this case, the IO event will be acceptable when
some of the task is done.
However, the mask won't change when the new mask is set. This makes the
dispatcher unable to be waken up correctly.

Fix: Update the waiting mask for each of the testing loop iteration