Bug fix: The file server hung after used up all the FDs
Trigger: When there's too many FDs is open by the program. The socket fd
will stop waking the event loop and deny any service

Cause: This is because of the nature of the edge trigger, this requires
for each epoll wakeup, we need to exhuasted all the connections from the
socket. But previously once the accept system call failed to return, it
will stop accepting any thing. However, the socket backlog might have
more connections, and this prevent the socket fd from waking the epoll
again.

Solution: We we realize there's an error, what we should do is use epoll
to sleep a little while.