Remove all the purposed RLS change
Since the reverse proxy is actually a different requirement than the
client. The difference is the graph *will not* use the result returned
by the server. Instead it directly return the result to client so it's
better to make this use case a proxy servlet which just return a RLS
token and the actual server access is delayed util the RLS has been
dereferenced by the module.

This requires the RLS deference interface supports event driven.