Bug fix: HTTP Request parser's routing table doesn't support prefix
This is very important, because we need to configure the server like
this:
	plumberserver.com/jenkins/ -> redirect to the jenkens instance
	plumberserver.com/         -> redirect to the static server

Thus we need to match the routing table with the longest prefix.

So we changed the sematics of the trie matcher a little bit.

First, we create a node whenever we see a word is fully matched.
Second, fix the bug: when we create tree, we create the child based on
the value of last bit. However, we when match in the tree, we actually
use the xor result to determine where to go. This is correct if we have
the both child, but will broken if we only have right child. So we
change the definition to if the last bit is the same, go to left child,
otherwise go to right. Thus if we only have one child, it has to be the
left child.
Third, we fix the error which crashes the program when we create a node
with single child
Fourth, we changed the matching function, when we see a full match we just
return and save the next state, instead of mark it as done.
Last, we changed the routing map behavior for the longest prefix match