Bug fix: TLS pipe might access freed memory
Repo: In the HTTP server, this happens when the Request Parser failed to
establish the TLS connection and sucessfully read the bytes. And for
some reason the servelt fails, which triggers the Response Render
renders a HTTP 500 response to the pipe. However, at the time the
response reader want to write the bytes, the read side of the pipe
handle has been disposed because the response parser is finished with
the failure

Fix: We actually check if the connection haven't been established even
after the input side of the pipe pair has been disposed.
If this is the case, it's an indicator for we are not able to establish
the the connection (Because we assume all the servlet must takes input
before move forward, thus if the connection is not established it must
be an error case). In this case we simply drop all the bytes wants to
write, because the communication channel is not ready.